Fixing the free percentage calculation for a link with multiple cores

+++ src/ca/bcit/drawing/Figure.java
@@ -32,7 +32,7 @@ public abstract class Figure {
 
 	protected abstract void drawOutline(GraphicsContext gc, Color color);
 
-	public abstract int getInfo();
+	public abstract double getInfo();
 
 	void drawOutline(GraphicsContext gc) {
 		drawOutline(gc, Color.BLACK);
